6425 | 5 files | M DumpSource2/schemas/client/CDOTA_Modifier_Lycan_Wolf_Bite_Lifesteal.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Lycan_Wolf_Bite_Lifesteal.h,  M game/bin/built_from_cl.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/18591065/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Lycan_Wolf_Bite_Lifesteal.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Lycan_Wolf_Bite_Lifesteal : public CDOTA_Buff
 {
 	float32 creep_penalty;
-	int32 lifesteal_percent;
-	int32 lifesteal_range;
+	float32 lifesteal_percent;
+	float32 lifesteal_range;
 	CUtlVector< CHandle< C_BaseEntity > > m_vecTargets;
 };
